+++ client/src/pages/_app.tsx
@@ -4,7 +4,7 @@ import type { AppProps } from 'next/app';
 import { useRouter } from 'next/router';
 
 import Head from 'next/head';
-import Header from '@/components/common/Header';
+import Header from '@/components/common/Header/Header';
 import LoginModal from '@/components/common/LoginModal';
 
 function TILApp({ Component, pageProps }: AppProps): JSX.Element {
